Added back compaction disabled option
This was removed because we add modified the vendor file directly before and thought
it was not required anymore.

It was indeed not required for reprocessing eth-mainnet but it chokes on eth-ropsten
on a given block. So added it back but without the need to modify the leveldb dependency.

Instead, put the settings that triggers compaction very high so it's does not trigger
compaction and disabled Eth code path that were doing manual compaction.