Fix Arbitrum transaction type numbers
This moves the tx type numbers for Arbitrum transactions into the valid range (less than 128).  This range is required so that UnmarshalBinary will work.  UnmarshalBinary looks at the first byte of the binary format, and treats the tx as a LegacyTx if the first byte is >127, and as the tx type otherwise.  So the previous Arbitrum types would cause errors in unmarshaling from binary format.