Fix pruner.go conflict with OpenStorageTrie params
Upstream in 8bbb16b70ef, cachingDB.OpenStorageTrie was modified to take
the address instead of the address hash as one of its parameters. In
pruner.go dumpRawTrieDescendants, which is a function we added, there
was a call to OpenStorageTrie that needed to be updated. Followed the
same pattern for getting the address from the address hash in
nodeIterator.step() to resolve.