Fix eth_feeHistory gasUsedRatio
The intent for eth_feeHistory for Arbitrum is to return the ratio of gas
used per second / the target gas per second. The logic for calculating
this gets the receipts for each tx in a block and gets the cumulative
gas used, but it was using the wrong hash to look up the receipts (it
should be the block hash and not the receipts hash).

Note: the current implementation returns 1.0 for gasUsedRatio for the
blocks at the beginning of the window specified by eth_feeHistory's
blockCount and newestBlock if the prior block has the same timestamp as
the first block in the window (ie time has not ticked over).