Use the VersionedArbitrumSigner in most contexts
In most places in the go-ethereum codebase where MakeSigner is clled, there is a
header or a blockContext availble from which one can easily derive an
arbosVersion.

In all of those places, providing an accurate arbosVersion makes the signer more
stable than the one returned from MakeSigner which uses a hard-coded
`params.MaxArbosVersionSupported`.

We have carefully documented the reasoning for why we think each of the contexts
where we left calls to MakeSigner should be fine to use an Arbitrum signer
initialized with the `MaxArbosVersionSupported` value in https://www.notion.so/arbitrum/MakeSiger-vs-VersionedArbitrumSigner-1c401a3f59f880eb84bded5e721e2647?pvs=4