gas dimension tracing: Expose L1GasUsed, CallExecutionGas explicitly
This PR exposes the GasUsed, L1GasUsed, and L2GasUsed from the
transaction reciept directly in the tracer response. It also changes
the tracers for CALL, CALLCODE, STATICCALL, DELEGATECALL, CREATE and
CREATE2 to explicitly return the gas consumed by the child execution
of the call when the call stack depth is increased by those opcodes.