gas dimension tracing: write disk in sync batches, access list tracer
This commit deletes the async vibe coded solution for writing
transactions to disk in favor of writing them to disk synchronously in
batches of 1000 instead. It also replaces calls out to the statedb in
favor of using the logger.AccessListTracer which maintains a duplicate
parallel copy of the access list but also stores it in a map for faster
lookup. Hopefully these efforts help performance of the live tracer to
stay synced to the tip of the chain :(