Add SyncProgressMap to ethclient.Client
This is so that there is a method on ethclient that is actually useful
for retrieving the Arbitrum node syncing information.

The eth_syncing RPC call returns the full map already. This change is
just to have a convenient method on the ethclient RPC wrapper. We were
using the SyncProgress method in a Nitro test and it was confusing
why it was returning empty progress; that test can use this new method.