Add a TxGasLimitHook
This hook can be used to separate out the logic that's currently overloading the
GasChargingHook. We need to limit the gas per transaction before the intrinsic
gas is subtracted from the gas remaining instead of after.

This way, if there isn't enough gas to cover all of the execution and the
intrinsic gas (both of which are components of L2 gas,) the system can detect it
and report and appropraite error while respecting the actual per-transaction
gas limit.