Use ubuntu-4 for GitHub actions
This switches the actions from the over-committed `ubuntu-latest` GitHub
standard machines to OCL's own runners.

In addition, this change also upgrades the versions of the checkout and setup-go
actions and switches to reading the version of go to use from the CI actions
into the go.mod file, and updtes the go.mod to match upstream.